+++ crates/store/src/db/sql/utils.rs
@@ -7,9 +7,8 @@ use miden_objects::{
     utils::Deserializable,
 };
 use rusqlite::{
-    params,
+    Connection, OptionalExtension, params,
     types::{Value, ValueRef},
-    Connection, OptionalExtension,
 };
 
 use crate::errors::DatabaseError;
@@ -38,28 +37,35 @@ pub fn schema_version(conn: &Connection) -> rusqlite::Result<usize> {
 
 /// Auxiliary macro which substitutes `$src` token by `$dst` expression.
 macro_rules! subst {
-    ($src:tt, $dst:expr) => {
+    ($src:tt, $dst:expr_2021) => {
         $dst
     };
 }
 
 /// Generates a simple insert SQL statement with parameters for the provided table name and fields.
+/// Supports optional conflict resolution (adding "| REPLACE" or "| IGNORE" at the end will generate
+/// "OR REPLACE" and "OR IGNORE", correspondingly).
 ///
 /// # Usage:
 ///
+/// ```ignore
+/// insert_sql!(users { id, first_name, last_name, age } | REPLACE);
 /// ```
-/// insert_sql!(users { id, first_name, last_name, age });
-/// ```
+///
 /// which generates:
-/// "INSERT INTO users (id, `first_name`, `last_name`, age) VALUES (?, ?, ?, ?)"
+/// ```sql
+/// INSERT OR REPLACE INTO `users` (`id`, `first_name`, `last_name`, `age`) VALUES (?, ?, ?, ?)
+/// ```
 macro_rules! insert_sql {
-    ($table:ident { $first_field:ident $(, $($field:ident),+)? $(,)? }) => {
+    ($table:ident { $first_field:ident $(, $($field:ident),+)? $(,)? } $(| $on_conflict:expr)?) => {
         concat!(
-            stringify!(INSERT INTO $table),
-            " (",
+            stringify!(INSERT $(OR $on_conflict)? INTO ),
+            "`",
+            stringify!($table),
+            "` (`",
             stringify!($first_field),
-            $($(concat!(", ", stringify!($field))),+ ,)?
-            ") VALUES (",
+            $($(concat!("`, `", stringify!($field))),+ ,)?
+            "`) VALUES (",
             subst!($first_field, "?"),
             $($(subst!($field, ", ?")),+ ,)?
             ")"