+++ crates/store/src/server/mod.rs
@@ -1,14 +1,19 @@
-use std::sync::Arc;
+use std::{sync::Arc, time::Duration};
 
 use miden_node_proto::generated::store::api_server;
-use miden_node_utils::errors::ApiError;
+use miden_node_utils::{errors::ApiError, tracing::grpc::store_trace_fn};
 use tokio::net::TcpListener;
 use tokio_stream::wrappers::TcpListenerStream;
+use tower_http::trace::TraceLayer;
 use tracing::info;
 
-use crate::{blocks::BlockStore, config::StoreConfig, db::Db, state::State, COMPONENT};
+use crate::{
+    COMPONENT, blocks::BlockStore, config::StoreConfig, db::Db,
+    server::db_maintenance::DbMaintenance, state::State,
+};
 
 mod api;
+mod db_maintenance;
 
 /// Represents an initialized store component where the RPC connection is open, but not yet actively
 /// responding to requests.
@@ -18,6 +23,7 @@ mod api;
 /// components.
 pub struct Store {
     api_service: api_server::ApiServer<api::StoreApi>,
+    db_maintenance_service: DbMaintenance,
     listener: TcpListener,
 }
 
@@ -40,6 +46,10 @@ impl Store {
                 .map_err(|err| ApiError::DatabaseConnectionFailed(err.to_string()))?,
         );
 
+        let db_maintenance_service = DbMaintenance::new(
+            Arc::clone(&state),
+            Duration::from_secs(config.db_optimization_interval_secs),
+        );
         let api_service = api_server::ApiServer::new(api::StoreApi { state });
 
         let addr = config
@@ -54,15 +64,21 @@ impl Store {
 
         info!(target: COMPONENT, "Database loaded");
 
-        Ok(Self { api_service, listener })
+        Ok(Self {
+            api_service,
+            db_maintenance_service,
+            listener,
+        })
     }
 
-    /// Serves the store's RPC API.
+    /// Serves the store's RPC API and DB maintenance background task.
     ///
     /// Note: this blocks until the server dies.
     pub async fn serve(self) -> Result<(), ApiError> {
+        tokio::spawn(self.db_maintenance_service.run());
+        // Build the gRPC server with the API service and trace layer.
         tonic::transport::Server::builder()
-            .trace_fn(miden_node_utils::tracing::grpc::store_trace_fn)
+            .layer(TraceLayer::new_for_grpc().make_span_with(store_trace_fn))
             .add_service(self.api_service)
             .serve_with_incoming(TcpListenerStream::new(self.listener))
             .await