+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancements
+
+- Implemented database optimization routine (#721).
+
 ### Fixes
 
 - Faucet webpage is missing `background.png` and `favicon.ico` (#672).
@@ -21,6 +25,10 @@
 - [BREAKING] Update `GetBlockInputs` RPC (#709).
 - [BREAKING] `CheckNullifiersByPrefix` now takes a starting block number (#707).
 - [BREAKING] Removed nullifiers from `SyncState` endpoint (#708).
+- [BREAKING] Added `batch_prover_url` to block producer configuration (#701).
+- [BREAKING] Added `block_prover_url` to block producer configuration (#719).
+- [BREAKING] Updated to Rust Edition 2024 (#727).
+- [BREAKING] MSRV bumped to 1.85 (#727).
 
 ### Enhancements